refactor: move routing logic to `http` package from `app`
