+++ internal/http/spies_test.go
@@ -1,4 +1,4 @@
-package app_test
+package http_test
 
 import "net/http"
